refactor: 컴포넌트와 페이지 분리 컴포넌트 안에 페이지가 있으니 복잡해 보여서 분리하였음.
