+++ src/App.tsx
@@ -1,6 +1,6 @@
-import CharacterInfoPage from './components/character-info/character-info-page'
-import LandingPage from './components/landing/landing-page'
-import StoryPage from './components/story/story-page'
+import CharacterInfoPage from './pages/character-info-page'
+import LandingPage from './pages/landing-page'
+import StoryPage from './pages/story-page'
 import { FullScrollPage } from './components/full-scroll-page'
 
 function App() {
